Updated types in xorec.cpp & xorec.hpp

+++ src/xorec/xorec.cpp
@@ -23,9 +23,9 @@ void xorec_init() {
 XorecResult xorec_encode(
   const uint8_t *XOREC_RESTRICT data_buffer,
   uint8_t *XOREC_RESTRICT parity_buffer,
-  uint32_t block_size,
-  uint32_t num_data_blocks,
-  uint32_t num_parity_blocks,
+  size_t block_size,
+  size_t num_data_blocks,
+  size_t num_parity_blocks,
   XorecVersion version
 ) {
   if (!XOREC_INIT_CALLED) throw_error("xorec_init() must be called before calling xorec_encode()");
@@ -35,7 +35,7 @@ XorecResult xorec_encode(
 
   std::memset(parity_buffer, 0, block_size * num_parity_blocks);
 
-  for (uint32_t i = 0; i < num_data_blocks; ++i) {
+  for (unsigned i = 0; i < num_data_blocks; ++i) {
     void * XOREC_RESTRICT parity_block = reinterpret_cast<void*>(parity_buffer + (i % num_parity_blocks) * block_size);
     const void * XOREC_RESTRICT data_block = reinterpret_cast<const void*>(data_buffer + i * block_size);
 
@@ -58,9 +58,9 @@ XorecResult xorec_encode(
 XorecResult xorec_decode(
   uint8_t *XOREC_RESTRICT data_buffer,
   const uint8_t *XOREC_RESTRICT parity_buffer,
-  uint32_t block_size,
-  uint32_t num_data_blocks,
-  uint32_t num_parity_blocks,
+  size_t block_size,
+  size_t num_data_blocks,
+  size_t num_parity_blocks,
   const uint8_t * XOREC_RESTRICT block_bitmap,
   XorecVersion version
 ) {
@@ -71,7 +71,7 @@ XorecResult xorec_decode(
 
   if (!is_recoverable(block_bitmap, num_data_blocks, num_parity_blocks)) return XorecResult::DecodeFailure;
 
-  for (uint32_t i = 0; i < num_data_blocks; ++i) {
+  for (unsigned i = 0; i < num_data_blocks; ++i) {
     if (block_bitmap[i]) continue;
 
     void * XOREC_RESTRICT recover_block = reinterpret_cast<void*>(data_buffer + i * block_size);
@@ -81,7 +81,7 @@ XorecResult xorec_decode(
     memcpy(recover_block, parity_block, block_size);
 
     // XOR the recover block with the other data blocks
-    for (uint32_t j = i % num_parity_blocks; j < num_data_blocks; j += num_parity_blocks) {
+    for (unsigned j = i % num_parity_blocks; j < num_data_blocks; j += num_parity_blocks) {
       if (i == j) continue;
 
       const void * XOREC_RESTRICT data_block = reinterpret_cast<const void*>(data_buffer + j * block_size);
@@ -106,37 +106,37 @@ XorecResult xorec_decode(
 XorecResult xorec_prefetch_encode(
   const uint8_t *XOREC_RESTRICT data_buffer,  // unified memory
   uint8_t *XOREC_RESTRICT parity_buffer,      // host memory
-  uint32_t block_size,
-  uint32_t num_data_blocks,
-  uint32_t num_parity_blocks,
-  uint32_t prefetch_bytes,
+  size_t block_size,
+  size_t num_data_blocks,
+  size_t num_parity_blocks,
+  size_t prefetch_bytes,
   XorecVersion version
 ) {
   
   if (!XOREC_INIT_CALLED) throw_error("xorec_init() must be called before calling xorec_pipelined_encode()");
   if (xorec_check_args(block_size, num_data_blocks, num_parity_blocks) != XorecResult::Success) return XorecResult::InvalidCounts;
 
-  uint32_t prefetch_blocks = prefetch_bytes / block_size;
+  size_t prefetch_blocks = prefetch_bytes / block_size;
   if (prefetch_blocks == 0) prefetch_blocks = 1;
 
   cudaStream_t prefetch_stream;
   cudaStreamCreate(&prefetch_stream);
 
 
-  uint32_t initial_prefetch_bytes = std::min(prefetch_blocks * block_size, num_data_blocks * block_size);
+  size_t initial_prefetch_bytes = std::min(prefetch_blocks * block_size, num_data_blocks * block_size);
   cudaMemPrefetchAsync(data_buffer, initial_prefetch_bytes, cudaCpuDeviceId, prefetch_stream);
   
   std::memset(parity_buffer, 0, block_size * num_parity_blocks);
 
 
 
-  for (uint32_t i = 0; i < num_data_blocks; ++i) {
+  for (unsigned i = 0; i < num_data_blocks; ++i) {
     if (i % prefetch_blocks == 0) { // We are at a prefetch interval
       cudaStreamSynchronize(prefetch_stream);
 
       if (i + prefetch_blocks < num_data_blocks) {
-        uint32_t remaining_blocks = num_data_blocks - (i + prefetch_blocks);
-        uint32_t prefetch_blks = std::min(prefetch_blocks, remaining_blocks);
+        size_t remaining_blocks = num_data_blocks - (i + prefetch_blocks);
+        size_t prefetch_blks = std::min(prefetch_blocks, remaining_blocks);
         cudaMemPrefetchAsync(data_buffer + (i+prefetch_blocks) * block_size, prefetch_blks * block_size, cudaCpuDeviceId, prefetch_stream);
       }
     }
@@ -168,11 +168,11 @@ XorecResult xorec_prefetch_encode(
 XorecResult xorec_prefetch_decode(
   uint8_t *XOREC_RESTRICT data_buffer,
   uint8_t *XOREC_RESTRICT parity_buffer,
-  uint32_t block_size,
-  uint32_t num_data_blocks,
-  uint32_t num_parity_blocks,
+  size_t block_size,
+  size_t num_data_blocks,
+  size_t num_parity_blocks,
   const uint8_t * XOREC_RESTRICT block_bitmap,
-  uint32_t prefetch_bytes,
+  size_t prefetch_bytes,
   XorecVersion version
 ) {
 
@@ -182,28 +182,28 @@ XorecResult xorec_prefetch_decode(
   if (xorec_check_args(block_size, num_data_blocks, num_parity_blocks) != XorecResult::Success) return XorecResult::InvalidCounts;
   if (!is_recoverable(block_bitmap, num_data_blocks, num_parity_blocks)) return XorecResult::DecodeFailure;
 
-  for (uint32_t i = 0; i < num_data_blocks; ++i) {
+  for (unsigned i = 0; i < num_data_blocks; ++i) {
     if (!block_bitmap[i]) restore_parity_bitmap[i % num_parity_blocks] = 1;
   }
 
 
-  uint32_t prefetch_blocks = prefetch_bytes / block_size;
+  size_t prefetch_blocks = prefetch_bytes / block_size;
 
   if (prefetch_blocks == 0) prefetch_blocks = 1;
 
   cudaStream_t prefetch_stream;
   cudaStreamCreate(&prefetch_stream);
 
-  uint32_t initial_prefetch_bytes = std::min(prefetch_blocks * block_size, num_data_blocks * block_size);
+  size_t initial_prefetch_bytes = std::min(prefetch_blocks * block_size, num_data_blocks * block_size);
   cudaMemPrefetchAsync(data_buffer, initial_prefetch_bytes, cudaCpuDeviceId, prefetch_stream);
 
-  for (uint32_t i = 0; i < num_data_blocks; ++i) {
+  for (unsigned i = 0; i < num_data_blocks; ++i) {
     if (i % prefetch_blocks == 0) { // We are at a prefetch interval
       cudaStreamSynchronize(prefetch_stream);
 
       if (i + prefetch_blocks < num_data_blocks) {
-        uint32_t remaining_blocks = num_data_blocks - (i + prefetch_blocks);
-        uint32_t prefetch_blks = std::min(prefetch_blocks, remaining_blocks);
+        size_t remaining_blocks = num_data_blocks - (i + prefetch_blocks);
+        size_t prefetch_blks = std::min(prefetch_blocks, remaining_blocks);
         cudaMemPrefetchAsync(data_buffer + (i+prefetch_blocks) * block_size, prefetch_blks * block_size, cudaCpuDeviceId, prefetch_stream);
       }
     }
@@ -226,7 +226,7 @@ XorecResult xorec_prefetch_decode(
     }
   }
 
-  for (uint32_t i = 0; i < num_data_blocks; ++i) {
+  for (unsigned i = 0; i < num_data_blocks; ++i) {
     if (!block_bitmap[i]) {
       memcpy(data_buffer + i * block_size, parity_buffer + (i % num_parity_blocks) * block_size, block_size);
     }
